+++ changelog.md
@@ -3,6 +3,7 @@
 ## Unreleased
 * Fix contact form placeholders generating invalid HTML
 * Email/Telephone fields in contact forms didn't get proper field type - fixed
+* Improve accessibility and styling of required field labels
 
 ## 3.4.1 (2020-04-16)
 * Updated German translations, courtesy of Wiljon Bolten